Add environment variables to the Open CLI specification
Environment variables can affect the way that a command-line tool
works in a variety of ways, as much as options, and arguments can.
Add a section to the specification where all of the environment
variables used by the tool can be declared and described.

Options can sometimes take their default value from an environment
variable too. Permit referring to an environment variable so that
this relationship can be documented.